// packages/@tuist/editor/src/lib/utils/types.ts
export type PathType = string[]

export interface ElementsType {
  [key: string]: ElementType
}

// Used during page preview

export interface ElementSize {
  // Content with padding.
  // content
  c: number
  // top margin
  t: number
  // bottom margin
  b: number
}

export interface LegacySize {
  // Legacy format
  content?: number
  topMargin?: number
  bottomMargin?: number
}

export interface ElementSizes {
  [key: string]: ElementSize
}

export interface PositionedElementType {
  p: number
}

export interface PositionedElementsType {
  [key: string]: PositionedElementType
}

export interface ElementOptionsType {
  // custom type name (used to get other options such as className, etc)
  t?: string
  // heading level
  h?: number
  // column attibution
  u?: 'l' | 'r'
  // column width
  uw?: number
  // list type
  l?: 'ol' | 'ul'
  href?: string
  // When true: export this paragraph.
  e?: boolean
  // When true: new page before this paragraph
  n?: boolean
  // align (justify, left, center, right)
  a?: 'j' | 'l' | 'c' | 'r'
  // When this is set, special behavior apply like
  // - cannot remove
  // - cannot change paragraph
  // - cannot set bold/etc (no toolbox)
  title?: boolean
  // For custom elements: open popup on selection
  open?: boolean
}

export interface ParagraphOptions {
  // Custom tag
  c?: string
  // Options to display paragraphs
  o?: ElementOptionsType
}

export interface GroupElementRefTypeById {
  [key: string]: GroupElementRefType
}

export interface GroupElementRefType {
  path: PathType
  elem: GroupElementType
}

export type Type = 'P' | 'A' | 'E' | 'T' | 'B' | 'I' | 'B+I'

export interface ElementType extends ParagraphOptions {
  // Type
  t: Type
  // Position
  p: number
  // Inner content (string)
  i?: string
  // Selection
  s?: SelectionType
  // Group
  g?: ElementsType
}
// ...
export interface ChangeType {
  op: 'noop' | 'update' | 'delete' | 'create'
  selected?: boolean
  path: PathType
  pathId: string
  elem: ElementType
  resized?: boolean
}

export interface ElementRefTypeById {
  [key: string]: ElementRefType
}

export interface StringElementRefType {
  path: PathType
  elem: StringElementType
}

export interface ElementNamedType {
  ref: string
  elem: ElementType
}

export interface CompositionType {
  sz?: ElementSizes

  toolbox?: ToolboxOperationValueType
  // Data for custom components.
  data?: { [key: string]: any }
  g: ElementsType
  // Path to object containing selection encoded as 'key.subKey.subsubKey'
  spath?: string
  // selection min paragraph position
  smin?: number
  // selection max paragraph position
  smax?: number
  // Title of the composition (when there is one)
  title?: string
}

// ============== Selections

export interface SelectionPositionType {
  top: number
  left: number
}

export interface CaretSelectionType {
  anchorPath: PathType
  anchorOffset: number
  anchorValue?: string
  position: SelectionPositionType
  stringPath?: string // Only used by setSelection
  type: 'Caret'
}

export interface RangeSelectionType {
  anchorPath: PathType
  anchorOffset: number
  anchorValue?: string
  focusPath: PathType
  focusOffset: number
  position: SelectionPositionType
  stringPath?: string // Only used by setSelection
  type: 'Range'
}

export type SelectionType = CaretSelectionType | RangeSelectionType

export interface ChangesType {
  // Initial data for custom elements (further changes do not go through
  // editor).
  data?: { [elemId: string]: any }
  // Full elements.
  elements: { [elemPath: string]: ChangeType }
  start: ChangeType
  end: ChangeType
  // Special cases where we need to prepare selection/toolbox during op eval.
  selection?: SelectionType
}
// ...
export interface SelectOperationType {
  op: 'select'
  value: SelectionType
}

export interface UpdateOperationType {
  op: 'update'
  path: PathType
  value: ElementType
}

export interface UpdateOptsOperationType {
  op: 'updateOpts'
  path: PathType
  opts: ParagraphOptions
}

export interface DataOperationType {
  op: 'data'
  path: PathType
  data: any
}

export interface DeleteOperationType {
  op: 'delete'
  path: PathType
}

export interface ResizedOperationType {
  op: 'resized'
  id: string
}

export interface ToolboxOpType {
  type: 'emptyParagraph' | 'paragraph' | 'selection'
  position: SelectionPositionType
}

interface ToolboxNoneType {
  type: 'none'
}

export type ToolboxOperationValueType = ToolboxOpType | ToolboxNoneType

export interface ToolboxOperationType {
  op: 'toolbox'
  value?: ToolboxOperationValueType
}

export type OperationType =
  | SelectOperationType
  | UpdateOperationType
  | UpdateOptsOperationType
  | DataOperationType
  | DeleteOperationType
  | ResizedOperationType
  | ToolboxOperationType